Refactors project with separated test sources
