+++ src/main/scala/Core.scala
@@ -1,9 +1,8 @@
-package org.apps.gist
-package university
+package org.hablapps.statesome
 
 import Function.const
 import scalaz._, Scalaz._
-import shapeless._, labelled._, shapeless.syntax.singleton._
+import shapeless._, labelled._
 
 /**
  * State algebras
@@ -203,161 +202,3 @@ object Util {
   }
 }
  
-import Util.{ Lens, _ }, AlgFunctor._, GetEvidence._
-
-
-/**
- * Data Model
- */
-
-object Primitive{
-  type IntegerP[P[_]]=Field[P,Int]
-  type BooleanP[P[_]]=Field[P,Boolean]
-  type StringP[P[_]]=Field[P,String]
-}
-
-import Primitive._
-
-case class Department[P[_],D](
-  self: Field[P,D],
-  budget: IntegerP[P])
-
-object Department {
-
-  def instance[P[_], D](implicit 
-      ge: GetEvidence[Department[P, D]]): Department[P, D] = 
-    ge.apply
-
-  implicit val DepartmentAlgFunctor = new AlgFunctor[Department] {
-    def amap[Q[_]: Functor, P[_]: Functor](f: Q ~> P) =
-      λ[Department[Q, ?] ~> Department[P, ?]] { alg =>
-        Department(alg.self.amap(f), alg.budget.amap(f))
-      }
-  }
-}
-
-case class University[P[_], U, D](
-  self: Field[P, U],
-  name: StringP[P],
-  math: Department[P, D])
-
-object University {
-
-  def instance[P[_], U, D](implicit 
-      ge: GetEvidence[University[P, U, D]]): University[P, U, D] =
-    ge.apply
-
-  implicit def UniversityAlgFunctor[D] = 
-    new AlgFunctor[University[?[_], ?, D]] {
-      def amap[Q[_]: Functor, P[_]: Functor](f: Q ~> P) =
-        λ[University[Q, ?, D] ~> University[P, ?, D]] { alg =>
-          University(alg.self.amap(f), alg.name.amap(f), alg.math.amap(f))
-        }
-    }
-}
-
-case class City[P[_], C, U, D](
-  self: Field[P, C],
-  popu: Field[P, Double],
-  univ: University[P, U, D])
-
-object City {
-  
-  def instance[P[_], C, U, D](implicit
-      ge: GetEvidence[City[P, C, U, D]]): City[P, C, U, D] =
-    ge.apply
-}
-
-case class Logic[P[_], U, D](univ: University[P, U, D]) {
-
-  def doubleBudget(implicit M: Monad[P]): P[Unit] =
-    univ.math.budget.modify(_ * 2)
-  
-  // should be reused
-  def getMathDep: P[D] = 
-    univ.math.self.get()
-}
-
-
-/** 
- * Data Layer Interpretation
- */
-
-case class SCity(population: Double, univ: SUniversity)
-
-object SCity {
- 
-  val popuLn: Lens[SCity, Double] =
-    Lens(_.population, p => _.copy(population = p))
-
-  val univLn: Lens[SCity, SUniversity] =
-    Lens(_.univ, u => _.copy(univ = u))
-}
-
-case class SUniversity(name: String, math: SDepartment)
-
-object SUniversity {
-
-  val nameLn: Lens[SUniversity, String] =
-    Lens(_.name, n => _.copy(name = n))
-  
-  val mathLn: Lens[SUniversity, SDepartment] =
-    Lens(_.math, d => _.copy(math = d))
-}
-
-case class SDepartment(budget: Int)
-
-object SDepartment {
-
-  val budgetLn = Lens[SDepartment, Int](_.budget, b => _.copy(budget = b))
-}
-
-import SCity._, SUniversity._, SDepartment._
-
-object Main extends App {
-
-  object TestDepartment {
-    implicit val self = 'self ->> lensId[SDepartment]
-    implicit val budg = 'budget ->> budgetLn
-
-    val StateDepartment =
-      Department.instance[State[SDepartment, ?], SDepartment]
-  }
-
-  object TestCity {
-    implicit val self = 'self ->> lensId[SCity]
-    implicit val popu = 'popu ->> popuLn
-    implicit val univ = 'univ ->> univLn
-    implicit val sel1 = 'self ->> (univ compose lensId[SUniversity])
-    implicit val name = 'name ->> (univ compose nameLn)
-    implicit val math = 'math ->> (univ compose mathLn)
-    implicit val sel2 = 'self ->> (math compose lensId[SDepartment])
-    implicit val budg = 'budget ->> (math compose budgetLn)
-
-    val StateCity = 
-      City.instance[State[SCity, ?], SCity, SUniversity, SDepartment]
-  }
-
-  // University example (includes some logic)
-  
-  implicit val sel1 = 'self ->> lensId[SUniversity]
-  implicit val name = 'name ->> nameLn
-  implicit val matx = 'math ->> mathLn
-  implicit val sel2 = 'self ->> (mathLn compose lensId[SDepartment])
-  implicit val budg = 'budget ->> (mathLn compose budgetLn)
-
-  val StateUniversity = 
-    University.instance[State[SUniversity, ?], SUniversity, SDepartment]
-
-  val logic = Logic(StateUniversity)
-  
-  val math = SDepartment(3000)
-  val univ = SUniversity("urjc", math)
-
-  val univ2 = logic.doubleBudget(implicitly).exec(univ)
-  val deps  = logic.getMathDep.eval(univ2)
-
-  println(s"univ2 = $univ2")
-  println(s"deps  = $deps")
-}
-
